Refactor chart styling utilities to improve color handling
- Replaced the deprecated getSeriesStyles function with separate getSeriesStroke and getSeriesLineStyles functions for better clarity and modularity.
- Updated components to utilize the new functions, ensuring that color resolution logic now prioritizes the overrideColor parameter for consistent styling across charts.
- Removed obsolete tests related to getSeriesStyles to streamline the testing suite.
