+++ projects/js-packages/charts/src/utils/get-styles.ts
@@ -1,7 +1,6 @@
 import { LineStyles } from '@visx/xychart';
 import { CSSProperties } from 'react';
 import { ChartTheme, SeriesData } from '../types';
-import type { ChartContextValue } from '../providers/chart-context';
 import type { LegendShape } from '@visx/legend/lib/types';
 
 /**
@@ -46,33 +45,6 @@ export function getSeriesStroke(
 ): string {
 	// Legacy fallback (when not in a ChartProvider). This is kept for non-context usages (e.g., tests).
 	return seriesData.options?.stroke ?? themeColors[ index % themeColors.length ];
-}
-
-/**
- * Combined utility that returns both stroke and line styles
- *
- * @param {SeriesData} seriesData        - The series data containing styling options
- * @param {number}     index             - The index of the series in the data array
- * @param {ChartTheme} providerTheme     - The chart theme configuration
- * @param              resolveGroupColor - Optional resolver from ChartContext for stable group colors
- * @return {object} Object containing stroke color and line styles
- */
-export function getSeriesStyles(
-	seriesData: SeriesData,
-	index: number,
-	providerTheme: ChartTheme,
-	resolveGroupColor?: ChartContextValue[ 'resolveGroupColor' ]
-): { stroke: string; lineStyles: LineStyles } {
-	const stroke = resolveGroupColor
-		? resolveGroupColor( {
-				group: seriesData.group,
-				index,
-				overrideColor: seriesData.options?.stroke,
-		  } )
-		: getSeriesStroke( seriesData, index, providerTheme.colors );
-
-	const lineStyles = getSeriesLineStyles( seriesData, index, providerTheme );
-	return { stroke, lineStyles };
 }
 
 /**
